+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 
+## [v1.3.0] - 2024-01-24
+### Changed
+- The package name from `egg_timer` to `eggtimer`
+- Deprecated the `egg_timer` package
+
 ## [v1.2.0] - 2023-03-28
 ### Added
 - ThreadSafeEggTimer